adjust data type to match CLM

+++ GEOSagcm_GridComp/GEOSphysics_GridComp/GEOSsurface_GridComp/GEOSland_GridComp/GEOScatchCN_GridComp/GEOScatchCNCLM51_GridComp/CLM51/CNCLM51_Photosynthesis.F90
@@ -121,7 +121,7 @@
 
  ! local variables to compute Photosynthesis inputs
  real                           :: ws, wl
- real, allocatable, dimension(:,:) :: rho, tau
+ real(r8), allocatable, dimension(:,:) :: rho, tau
  real, dimension (nch, NUM_ZON) :: esat_tv     ! vapor pressure inside leaf (sat vapor press at tc) (Pa)
  real, dimension (nch, NUM_ZON) :: eair        ! vapor pressure of canopy air
  real, dimension (nch)          :: oair        ! Atmospheric O2 partial pressure (Pa)
